+++ components/application/stages/StagePre.tsx
@@ -10,8 +10,8 @@ export default function StagePre() {
     <div className="flex w-9/10 flex-col gap-6 rounded-lg bg-gray-1 p-8 sm:w-[clamp(400px,50%,500px)]">
       <QuestionsContextProvider
         questions={[
-          <QuestionSeekingRomance key="seekingRomance" />,
           <QuestionIncarcerated key="incarcerated" />,
+          <QuestionSeekingRomance key="seekingRomance" />,
           <QuestionSuccess key="success" />,
         ]}
       >
